requirements.txt
faiss-gpu
GitPython
hugchat
huggingface_hub
langchain
langchain_community
langchain_core
langchain_huggingface
langchain_text_splitters
torch  --index-url https://download.pytorch.org/whl/cu124
torchvision  --index-url https://download.pytorch.org/whl/cu124
torchaudio  --index-url https://download.pytorch.org/whl/cu124
numpy
diffusers
accelerate
sentence_transformers
xformers
llama-cpp-python
pandas
playsound
plotly
python-dotenv
pyttsx3
Requests
rich
scikit_learn
scipy
streamlit
streamlit_option_menu
transformers
TTS
wandb